--- approval-common/src/main/java/com/sast/approval/Exception/BaseException.java
@@ -1,14 +1,0 @@
-package com.sast.approval.Exception;
-
-import com.sast.approval.enums.ErrorEnum;
-import lombok.Getter;
-
-@Getter
-public class BaseException extends RuntimeException {
-    private final ErrorEnum errorEnum;
-    public BaseException(ErrorEnum errorEnum) {
-        super(errorEnum.getErrMsg());
-        this.errorEnum = errorEnum;
-    }
-
-}
