--- approval-common/src/main/java/com/sast/approval/utils/RedisUtil.java
@@ -1,7 +1,0 @@
-package com.sast.approval.utils;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class RedisUtil {
-}
