+++ approval-common/src/main/java/fun/sast/Exception/BaseException.java
@@ -1,0 +1,14 @@
+package fun.sast.Exception;
+
+import fun.sast.enums.ErrorEnum;
+import lombok.Getter;
+
+@Getter
+public class BaseException extends RuntimeException {
+    private final ErrorEnum errorEnum;
+    public BaseException(ErrorEnum errorEnum) {
+        super(errorEnum.getErrMsg());
+        this.errorEnum = errorEnum;
+    }
+
+}
