+++ approval-common/src/main/java/fun/sast/utils/RedisUtil.java
@@ -1,0 +1,7 @@
+package fun.sast.utils;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class RedisUtil {
+}
