--- approval-pojo/src/main/java/com/sast/approval/entity/Department.java
@@ -1,30 +1,0 @@
-package com.sast.approval.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-import java.time.LocalDateTime;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Department implements Serializable {
-    /**
-     * 部门id
-     */
-    private Integer id;
-    /**
-     * 部门名称
-     */
-    private String name;
-
-    private LocalDateTime createTime;
-
-    private LocalDateTime updateTime;
-
-    private Long createUser;
-
-    private Long updateUser;
-}
