+++ approval-pojo/src/main/java/fun/sast/dto/UserLoginDTO.java
@@ -1,0 +1,21 @@
+package fun.sast.dto;
+
+import lombok.Data;
+
+import java.io.Serializable;
+
+@Data
+public class UserLoginDTO implements Serializable {
+    /**
+     * 登录验证码
+     */
+    private String validateCode;
+    /**
+     * 登录密码
+     */
+    private String password;
+    /**
+     * 学号
+     */
+    private String code;
+}
