--- approval-server/src/main/java/com/sast/approval/ApprovalApplication.java
@@ -1,15 +1,0 @@
-package com.sast.approval;
-
-import lombok.extern.slf4j.Slf4j;
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-@Slf4j
-@SpringBootApplication
-@MapperScan("com.sast.approval.mapper")
-public class ApprovalApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(ApprovalApplication.class, args);
-        log.info("server started");
-    }
-}
