--- approval-server/src/main/java/com/sast/approval/annotation/ResponseResult.java
@@ -1,13 +1,0 @@
-package com.sast.approval.annotation;
-
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(java.lang.annotation.RetentionPolicy.RUNTIME)
-@Documented
-public @interface ResponseResult {
-}
