--- approval-server/src/main/java/com/sast/approval/mapper/UserMapper.java
@@ -1,9 +1,0 @@
-package com.sast.approval.mapper;
-
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.sast.approval.entity.User;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserMapper extends BaseMapper<User> {
-}
