--- approval-server/src/main/java/com/sast/approval/service/impl/UserServiceImpl.java
@@ -1,30 +1,0 @@
-package com.sast.approval.service.impl;
-
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.sast.approval.entity.User;
-import com.sast.approval.mapper.UserMapper;
-import com.sast.approval.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserServiceImpl implements UserService {
-
-    @Autowired
-    private UserMapper userMapper;
-
-    /**
-     * 验证用户信息
-     * @param code 学号
-     * @param password 密码
-     * @return 用户信息
-     */
-    @Override
-    public User authenticate(String code, String password) {
-//        QueryWrapper<User> queryWrapper = new QueryWrapper<>();
-//        queryWrapper.eq("code", code).eq("password", password);
-//        return userMapper.selectOne(queryWrapper);
-        return new User();
-        }
-
-}
