+++ approval-server/src/main/java/fun/sast/ApprovalApplication.java
@@ -1,0 +1,15 @@
+package fun.sast;
+
+import lombok.extern.slf4j.Slf4j;
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+@Slf4j
+@SpringBootApplication
+@MapperScan("fun.sast.mapper")
+public class ApprovalApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ApprovalApplication.class, args);
+        log.info("server started");
+    }
+}
