+++ approval-server/src/main/java/fun/sast/annotation/ResponseResult.java
@@ -1,0 +1,13 @@
+package fun.sast.annotation;
+
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(java.lang.annotation.RetentionPolicy.RUNTIME)
+@Documented
+public @interface ResponseResult {
+}
